# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lead Management CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dialer/sessions/{id}:
    get:
      operationId: getDialerSession
      tags: [leads]
      summary: Get dialer session by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DialerSession"
    patch:
      operationId: endDialerSession
      tags: [leads]
      summary: End a dialer session and save stats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DialerSessionUpdate"
      responses:
        "200":
          description: Session ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DialerSession"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, name, email, role, isBlocked, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, agent]
        isBlocked:
          type: boolean
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DialerSession:
      type: object
      required: [id, agentId, startedAt, totalCalls, connectedCalls, followUpsScheduled, dealsWon, revenueGenerated]
      properties:
        id:
          type: integer
        agentId:
          type: integer
        startedAt:
          type: string
        endedAt:
          type: ["string", "null"]
        totalCalls:
          type: integer
        connectedCalls:
          type: integer
        followUpsScheduled:
          type: integer
        dealsWon:
          type: integer
        revenueGenerated:
          type: number
    DialerSessionUpdate:
      type: object
      properties:
        totalCalls:
          type: integer
        connectedCalls:
          type: integer
        followUpsScheduled:
          type: integer
        dealsWon:
          type: integer
        revenueGenerated:
          type: number
...